add response channel
For every incoming task, the gateway now returns both a request
and a response sender. A response sender could be a Sender of a
channel (in the case of FileGateway) or a TcpStream (in the case
of HttpGateway). Syntax is similar to passing a socket fd to
a pthread in multi-threaded web servers.