Prototype very basic protobuf based syscalls
This commit adds a protocol buffer description in `src/syscalls.proto`
and includes some basic code for actually using it from the VMM in
`src/vm.rs` in a new and unused method `process_syscall`.

The supported system calls are:

- `ReadKey` and
- `WriteKey`

which use an lmdb database to read and write keys (though this is
totally untested)

and `Response` which is meant to replace how apps currently submit their
responses to requests but is unhandled and untested (the theme is that
nothing in this commit is tested).

The basic format of the RPC protocol is to read a 4-byte big-endian u32
from the vsock which encodes the lenght of the system call request,
decode the system call request, the respond in the same manner (4-byte
big-endian u32 length followed by the response message).