Avoid busy looping in gateway
Rather than busy loop on all connections, just spawn a new thread for
each connection.

We don't anticipate many concurrent connections for the gateway (one per
front-end), so it's not a scalability bottleneck anyway. And if it
becomes one, we can/should use a non-blocking event system instead of
busy looping anyway.