Add CLI writing/reading blobs
Adds a command line interface, `sfblob` for reading/writing blobs. For
writing, it takes input from stdin, stores it in the blobstore and
outputs the blob name to stdout. For reading, it takes one or more lines
of blob names on stdin and outputs the named blobs' contents to stdout.